+++ src/extension.ts
@@ -1,5 +1,9 @@
+// The module 'vscode' contains the VSCode extensibility API
+// Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
+import { VSCodeMarketplaceClient } from 'vscode-marketplace-client';
 
+// Import the Configs, Controllers, and Providers
 import {
   Config,
   EXTENSION_DISPLAY_NAME,
@@ -79,12 +83,13 @@ export async function activate(context: vscode.ExtensionContext) {
       config.update(workspaceConfig);
 
       if (isEnabled) {
-        const message = vscode.l10n.t('{0} is now enabled and ready to use', [
-          EXTENSION_DISPLAY_NAME,
-        ]);
+        const message = vscode.l10n.t(
+          'The {0} extension is now enabled and ready to use',
+          [EXTENSION_DISPLAY_NAME],
+        );
         vscode.window.showInformationMessage(message);
       } else {
-        const message = vscode.l10n.t('{0} is now disabled', [
+        const message = vscode.l10n.t('The {0} extension is now disabled', [
           EXTENSION_DISPLAY_NAME,
         ]);
         vscode.window.showInformationMessage(message);
@@ -124,12 +129,12 @@ export async function activate(context: vscode.ExtensionContext) {
         title: vscode.l10n.t('Release Notes'),
       },
       {
-        title: vscode.l10n.t('Close'),
+        title: vscode.l10n.t('Dismiss'),
       },
     ];
 
     const message = vscode.l10n.t(
-      'New version of {0} is available. Check out the release notes for version {1}',
+      "The {0} extension has been updated. Check out what's new in version {1}",
       [EXTENSION_DISPLAY_NAME, currentVersion],
     );
     vscode.window.showInformationMessage(message, ...actions).then((option) => {
@@ -154,6 +159,58 @@ export async function activate(context: vscode.ExtensionContext) {
 
     // Update the version in the global state
     context.globalState.update('version', currentVersion);
+  }
+
+  // -----------------------------------------------------------------
+  // Check for updates
+  // -----------------------------------------------------------------
+
+  // Check for updates to the extension
+  try {
+    // Retrieve the latest version
+    VSCodeMarketplaceClient.getLatestVersion(
+      USER_PUBLISHER,
+      EXTENSION_NAME,
+    ).then((latestVersion) => {
+      // Check if the latest version is different from the current version
+      if (latestVersion !== currentVersion) {
+        const actions: vscode.MessageItem[] = [
+          {
+            title: vscode.l10n.t('Update Now'),
+          },
+          {
+            title: vscode.l10n.t('Dismiss'),
+          },
+        ];
+
+        const message = vscode.l10n.t(
+          'A new version of {0} is available. Update to version {1} now',
+          [EXTENSION_DISPLAY_NAME, latestVersion],
+        );
+        vscode.window
+          .showInformationMessage(message, ...actions)
+          .then(async (option) => {
+            if (!option) {
+              return;
+            }
+
+            // Handle the actions
+            switch (option?.title) {
+              case actions[0].title:
+                await vscode.commands.executeCommand(
+                  'workbench.extensions.action.install.anotherVersion',
+                  `${USER_PUBLISHER}.${EXTENSION_NAME}`,
+                );
+                break;
+
+              default:
+                break;
+            }
+          });
+      }
+    });
+  } catch (error) {
+    console.error('Error retrieving extension version:', error);
   }
 
   // -----------------------------------------------------------------
@@ -339,7 +396,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -355,7 +412,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -371,7 +428,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -387,7 +444,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -403,7 +460,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -419,7 +476,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -435,7 +492,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -451,7 +508,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -467,7 +524,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -483,7 +540,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -499,7 +556,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -515,7 +572,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -531,7 +588,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -547,7 +604,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -563,7 +620,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -579,7 +636,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -595,7 +652,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -611,7 +668,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -627,7 +684,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -643,7 +700,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -659,7 +716,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -675,7 +732,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -699,7 +756,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -715,7 +772,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -731,7 +788,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -747,7 +804,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -763,7 +820,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -779,7 +836,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -795,7 +852,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -811,7 +868,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -827,7 +884,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -843,7 +900,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -859,7 +916,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -875,7 +932,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -891,7 +948,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -907,7 +964,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -931,7 +988,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -955,7 +1012,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -972,7 +1029,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -1005,7 +1062,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -1038,7 +1095,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -1078,7 +1135,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -1118,7 +1175,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);
@@ -1151,7 +1208,7 @@ export async function activate(context: vscode.ExtensionContext) {
       // Check if the extension is enabled
       if (!config.enable) {
         const message = vscode.l10n.t(
-          '{0} is disabled in settings. Enable it to use its features',
+          'The {0} extension is disabled in settings. Enable it to use its features',
           [EXTENSION_DISPLAY_NAME],
         );
         vscode.window.showErrorMessage(message);